Make it in a separate command, log to stderr, use config as default
